refactor(chat): 提取历史窗口装配
将规范 JSONL 与展示快照的窗口坐标、分页和 around UUID 装配移入独立模块，展示叠加层复用该实现，chat.py 保留兼容门面与既有路由响应契约。
