+++ backend/chat_overlays.py
@@ -19,6 +19,7 @@ import time
 from typing import Any, Callable
 import uuid
 
+from . import chat_history_window
 from . import sessions as sess
 from . import transcript_index as transcript_idx
 from .settings import atomic_write_text
@@ -1089,9 +1090,7 @@ def _load_cancelled_turn_snapshots(sid: str) -> tuple[list[dict], str]:
 
 
 def _combined_history_generation(base: str, snapshot_generation: str) -> str:
-    if not snapshot_generation:
-        return base
-    return f"{base or 'none'}~cancelled-{snapshot_generation}"
+    return chat_history_window.combined_generation(base, snapshot_generation)
 
 
 def _persist_cancelled_turn_snapshot(bc: "TurnBroadcast") -> bool:
@@ -1480,88 +1479,7 @@ def _interrupted_history_segments(
     snapshots: list[dict],
     order: str,
 ) -> tuple[list[dict], int]:
-    """Build a virtual bubble stream of canonical records + snapshots."""
-    records = (index or {}).get("records") or []
-    orders = (index or {}).get("orders") or {}
-    record_ids = list(orders.get(order) or [])
-    positions: dict[str, int] = {}
-    for position, record_id in enumerate(record_ids):
-        uid = str(records[record_id].get("uuid") or "")
-        if uid:
-            positions[uid] = position
-    full_uuids = {
-        str(records[record_id].get("uuid") or "")
-        for record_id in orders.get("full") or []
-    }
-    prefix = ((index or {}).get("bubble_prefix") or {}).get(order) or [0]
-
-    placed: dict[int, list[dict]] = {}
-    included: list[dict] = []
-    for snapshot in snapshots:
-        anchors = snapshot.get("anchors") or {}
-        anchor = anchors.get(order) or {}
-        anchor_uuid = str(anchor.get("uuid") or "")
-        if anchor_uuid and anchor_uuid in positions:
-            position = positions[anchor_uuid]
-        elif order == "normal" and anchor_uuid and anchor_uuid in full_uuids:
-            # The interrupted turn sits before the current compact/fork root.
-            # Keep it in full-history mode but do not punch it back into the
-            # normal post-compact view.
-            continue
-        else:
-            target = max(0, int(anchor.get("total") or 0))
-            position = -1
-            for pos in range(len(record_ids)):
-                reached = int(prefix[pos + 1] if pos + 1 < len(prefix) else 0)
-                if reached > target:
-                    break
-                position = pos
-        placed.setdefault(position, []).append(snapshot)
-        included.append(snapshot)
-
-    for group in placed.values():
-        group.sort(key=lambda item: (
-            int(item.get("started_at_ms") or 0), str(item.get("turn_id") or "")))
-
-    hidden = {
-        str(uid)
-        for snapshot in included
-        for uid in (snapshot.get("hidden_uuids") or [])
-        if uid
-    }
-    segments: list[dict] = []
-
-    def append_snapshots(position: int) -> None:
-        for snapshot in placed.get(position, []):
-            messages = snapshot.get("messages") or []
-            if messages:
-                segments.append({
-                    "kind": "snapshot",
-                    "snapshot": snapshot,
-                    "count": len(messages),
-                })
-
-    append_snapshots(-1)
-    for position, record_id in enumerate(record_ids):
-        record = records[record_id]
-        if str(record.get("uuid") or "") not in hidden:
-            count = max(0, int(record.get("bubble_count") or 0))
-            if count:
-                segments.append({
-                    "kind": "record",
-                    "record_id": record_id,
-                    "count": count,
-                })
-        append_snapshots(position)
-    snapshot_turns = sum(
-        1
-        for snapshot in included
-        if any(
-            isinstance(message, dict) and message.get("role") == "user"
-            for message in (snapshot.get("messages") or [])
-        )
-    )
-    return segments, snapshot_turns
+    return chat_history_window.history_segments(index, snapshots, order)
 
 
 def _interrupted_history_stats(
@@ -1569,17 +1487,7 @@ def _interrupted_history_stats(
     snapshots: list[dict],
     order: str,
 ) -> tuple[int, int]:
-    segments, snapshot_turns = _interrupted_history_segments(
-        index, snapshots, order)
-    total = sum(int(segment.get("count") or 0) for segment in segments)
-    records = (index or {}).get("records") or []
-    canonical_turns = sum(
-        1
-        for segment in segments
-        if segment.get("kind") == "record"
-        and records[segment["record_id"]].get("real_user_prompt")
-    )
-    return total, canonical_turns + snapshot_turns
+    return chat_history_window.history_stats(index, snapshots, order)
 
 
 def _interrupted_history_window(
@@ -1593,75 +1501,17 @@ def _interrupted_history_window(
     offset: int = -1,
     limit: int = 0,
 ) -> tuple[list[dict], int, int, bool]:
-    """Read one window from the virtual display history.
-
-    Only canonical records that intersect the requested display window are
-    read from JSONL.  A session with a cancelled snapshot therefore retains
-    the existing long-history/windowing protections instead of falling back to
-    a full transcript parse on every open.
-    """
-    _hooks = _require_hooks()
-    segments, _ = _interrupted_history_segments(index, snapshots, order)
-    total = sum(int(segment.get("count") or 0) for segment in segments)
-    if offset >= 0:
-        start = max(0, min(offset, total))
-        end = total if limit <= 0 else min(total, start + limit)
-    elif tail > 0:
-        start = max(0, total - tail)
-        end = total
-    else:
-        start, end = 0, total
-
-    pieces: list[tuple[str, Any, int, int]] = []
-    selected_record_ids: list[int] = []
-    cursor = 0
-    for segment in segments:
-        count = int(segment.get("count") or 0)
-        seg_start, seg_end = cursor, cursor + count
-        cursor = seg_end
-        overlap_start = max(start, seg_start)
-        overlap_end = min(end, seg_end)
-        if overlap_start >= overlap_end:
-            continue
-        local_start = overlap_start - seg_start
-        local_end = overlap_end - seg_start
-        if segment["kind"] == "record":
-            record_id = int(segment["record_id"])
-            selected_record_ids.append(record_id)
-            pieces.append(("record", record_id, local_start, local_end))
-        else:
-            pieces.append((
-                "snapshot", segment["snapshot"], local_start, local_end))
-
-    shaped_by_record: dict[int, list[dict]] = {}
-    if transcript_path is not None and index is not None and selected_record_ids:
-        shaped = _hooks.indexed_ui_records(
-            transcript_path, index, selected_record_ids, annotations)
-        shaped_cursor = 0
-        records = index.get("records") or []
-        for record_id in selected_record_ids:
-            count = max(0, int(records[record_id].get("bubble_count") or 0))
-            shaped_by_record[record_id] = shaped[
-                shaped_cursor:shaped_cursor + count]
-            shaped_cursor += count
-
-    window: list[dict] = []
-    for kind, source, local_start, local_end in pieces:
-        if kind == "record":
-            messages = shaped_by_record.get(int(source), [])
-        else:
-            started_at_ms = int(source.get("started_at_ms") or 0)
-            turn_id = str(source.get("turn_id") or "")
-            messages = [dict(message)
-                        for message in (source.get("messages") or [])]
-            for message in messages:
-                if turn_id:
-                    message.setdefault(
-                        "_turn_origin_uuid", f"terminal:{turn_id}")
-                if started_at_ms > 0:
-                    message.setdefault("turn_started_at", started_at_ms)
-        window.extend(messages[local_start:local_end])
-    return window, total, start, end < total
+    return chat_history_window.history_window(
+        transcript_path,
+        index,
+        snapshots,
+        annotations,
+        order,
+        shape_records=_require_hooks().indexed_ui_records,
+        tail=tail,
+        offset=offset,
+        limit=limit,
+    )
 
 
 def _interrupted_history_window_around_uuid(
@@ -1675,52 +1525,14 @@ def _interrupted_history_window_around_uuid(
     *,
     limit: int = 0,
 ) -> tuple[list[dict], int, int, bool] | None:
-    """Read an around-UUID window in the same virtual coordinates as paging."""
-    segments, _ = _interrupted_history_segments(index, snapshots, "full")
-    records = index.get("records") or []
-    total = sum(int(segment.get("count") or 0) for segment in segments)
-    cursor = 0
-    target_start = -1
-    target_end = -1
-    for segment in segments:
-        count = max(0, int(segment.get("count") or 0))
-        if segment.get("kind") == "record":
-            record_id = int(segment.get("record_id") or 0)
-            if (
-                0 <= record_id < len(records)
-                and str(records[record_id].get("uuid") or "") == uuid_value
-            ):
-                target_start = cursor
-                target_end = cursor + count
-                break
-        cursor += count
-    if target_start < 0 or target_end <= target_start:
-        return None
-
-    if limit > 0:
-        span = min(limit, target_end - target_start)
-        context = max(0, limit - span)
-        before_budget = context // 2
-        after_budget = context - before_budget
-        start = max(0, target_start - before_budget)
-        end = min(total, target_start + span + after_budget)
-        if end - start < limit:
-            start = max(0, end - limit)
-            end = min(total, start + limit)
-        if not (start <= target_start < end):
-            start = max(0, min(target_start, total - limit))
-            end = min(total, start + limit)
-    else:
-        start = max(0, target_start - max(0, before))
-        end = min(total, target_end + max(0, after))
-
-    window, _total, win_offset, has_later = _interrupted_history_window(
+    return chat_history_window.history_window_around_uuid(
         transcript_path,
         index,
         snapshots,
         annotations,
-        "full",
-        offset=start,
-        limit=end - start,
+        uuid_value,
+        before,
+        after,
+        shape_records=_require_hooks().indexed_ui_records,
+        limit=limit,
     )
-    return window, _total, win_offset, has_later
